make test run parallel

+++ __tests__/api/domains/domains.test.ts
@@ -3,6 +3,7 @@ import { execSync } from "child_process";
 import path from "path";
 import { PrismaClient } from "@prisma/client";
 import { getPublicKey } from "nostr-tools";
+import crypto from "crypto";
 
 import {
   GET,
@@ -13,6 +14,7 @@ import {
 import { POST as POST_VERIFY } from "../../../src/app/api/domains/[domain]/(domains)/verify/route";
 
 let prisma: PrismaClient;
+let testDbUrl: string;
 
 const RANDOM_PRIV_KEY =
   "3a7bd3e2360a3d29eea436fcfb7e44c735d117c42d1c1835420b6b9942dd4f1b";
@@ -20,9 +22,12 @@ const RANDOM_PRIV_KEY =
 const RANDOM_PUB_KEY =
   "92763cc6af957acc8159c3d0fbcd9f00e20b4222c1dcff07107190ff5f3667d8";
 
-const testDbUrl = `${path.join(__dirname, "../../../prisma/test.db")}`;
-
 beforeAll(async () => {
+  // Generate a unique database file name for each test run
+  const randomString = crypto.randomBytes(8).toString("hex");
+  const uniqueDbName = `test-${randomString}.db`;
+  testDbUrl = `${path.join(__dirname, "../../../prisma/", uniqueDbName)}`;
+
   // Set up the test database
   process.env.DATABASE_URL = `file:${testDbUrl}`;
 
@@ -35,10 +40,14 @@ beforeAll(async () => {
 });
 
 afterAll(async () => {
+  await prisma.$disconnect();
+
   try {
-    execSync(`rm ${testDbUrl}`, {
-      stdio: "inherit",
-    });
+    if (testDbUrl) {
+      execSync(`rm ${testDbUrl}`, {
+        stdio: "inherit",
+      });
+    }
   } catch (error) {
     console.warn(`Warning: Could not remove test database file: ${error}`);
   }
